Change Attributes collections param to an array
Latest Waterline master passes the collections to Schema() in as an
array. Schema() is documented to take the collections in as an array,
too. Updating Attributes() to match.

The original code actually almost works, unless some library
monkey-patches Array to add a method to it. The Attributes tries to
process the function as a collection, and hilarity ensues.