decide on how to handle identity case
This was all kinds of jumbled up. Going forward case shouldn’t be touched when setting or building any
mapping. The only time an identity is lowercased is when it’s stored on the collections map. This allows
normalized lookup usage. So whenever a collection is being looked up the identity should be lowercased.

The other issue was the difference between tableName and identity in association mappings. This clears that
up by including both.