Allow non-standard attribute properties
Adds a flag: attributeWhitelist to model to bypass attribute property validation. While this is fine, it doesn't allow for custom properties to be added to attributes.  This might be helpful to some users, but others might want to make distinct model schemas the contain isMobilePhone, or JSON Schema style snytax for an enum, etc.